+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+## [next]
+ - BodyComponent should follow Forge2D game debug mode
+ 
 ## [0.6.0-rc2]
  - Align with Flame 1.0.0-rc5